added date search
added form to get quakes from requested date range. the API limits the
number of quakes it can return, so i'm making a new request for each
new set of filters and going the traditional submit route to limit the
requests, rather than fetching each time an input is updated.